refactoring of the benchmark dataset capabilities and added a random txt slices dataset option
